+++ packages/service-library/src/web-api/api-contracts/privacy-scan-types/privacy-scan-run-request.ts
@@ -1,8 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+import { PrivacyScanConfiguration, StorageDocument } from 'storage-documents';
 import { ReportGroup, Website } from '../scan-run-request';
-import { PrivacyScanConfiguration } from './privacy-scan-configuration';
+
+export type PrivacyScanConfigDefinition = Omit<PrivacyScanConfiguration, keyof StorageDocument>;
 
 export interface PrivacyScanRequest {
     url: string;
