+++ packages/storage-documents/package.json
@@ -37,7 +37,6 @@
         "axe-result-converter": "^1.0.0",
         "common": "^1.0.0",
         "inversify": "^6.0.1",
-        "reflect-metadata": "^0.1.13",
-        "service-library": "^1.0.0"
+        "reflect-metadata": "^0.1.13"
     }
 }
