+++ packages/storage-documents/src/privacy-scan-types/privacy-scan-configuration.ts
@@ -1,11 +1,35 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { PrivacyScanConfiguration } from 'service-library';
-import { ItemType } from '..';
-import { StorageDocument } from '../storage-document';
+import { ItemType, StorageDocument } from '..';
 
-export type PrivacyScanConfigurationDocument = PrivacyScanConfiguration &
-    StorageDocument & {
-        itemType: ItemType.privacyScanConfiguration;
-    };
+// Data type stored by WCP
+export interface PrivacyScanConfiguration extends StorageDocument {
+    itemType: ItemType.privacyScanConfiguration;
+    ScanId: string;
+    Urls: string[];
+    Headers: ScanConfigHeaders;
+    Cookies: ScanConfigCookies;
+    ConsentCookiesPerRequest: ScanConfigCookies[];
+    MultiLoadTestCookies: ScanConfigCookies[];
+    DurationLimitMinutes: number;
+    Affinity: string;
+    CookieCollectionScan: boolean;
+    Whitelist: WhitelistDomain[];
+    BannerDetectionXPath: string;
+    ConsentModalDetection: ConsentModalDetection;
+}
+
+export type ScanConfigHeaders = { [key: string]: string };
+
+export type ScanConfigCookies = { [key: string]: string };
+
+export interface WhitelistDomain {
+    TopLevelDomain: string;
+    Whitelisted: boolean;
+}
+
+export interface ConsentModalDetection {
+    ManagementElementXPath: string;
+    ConsentModalXPath: string;
+}
